fix float

+++ bcl2fastq_pipeline/afterFastq.py
@@ -94,7 +94,7 @@ def fastq_screen_worker(fname) :
     ofile=fname.replace("_R1_001.fastq.gz","subsampled.fastq")
     #print("ofile: {}".format(ofile))
     wc = subprocess.check_output("zcat {} | wc -l ".format(fname), shell=True)
-    seqtk_size = int(int(wc)*int(config.get("fastq_screen","seqtk_fraction")))
+    seqtk_size = int(int(wc)*float(config.get("fastq_screen","seqtk_fraction")))
     cmd = "%s sample %s %s %s" % (
         config.get("fastq_screen","seqtk_command"),
         config.get("fastq_screen","seqtk_options"),
